Fixed _destroyed so it correctly removes model.ids
Previously _destroyed was trying to remove _idOf(model), which is
returning this.name + model.id. However from what I can tell this.records is
just the model.ids, non of which have this.name concated onto the front.
Hence the id names were not being removed from this.records.

I have also updated the minified version with this change.